backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
chromadb==0.4.15
sentence-transformers==2.2.2
together==0.2.7
pydantic==2.5.0
httpx==0.25.2